+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/api/business/GetBusinessAccountGiftsAction.kt
@@ -33,8 +33,10 @@ class GetBusinessAccountGiftsAction(
  * @param excludeUnsaved Pass True to exclude gifts that aren't saved to the account's profile page
  * @param excludeSaved Pass True to exclude gifts that are saved to the account's profile page
  * @param excludeUnlimited Pass True to exclude gifts that can be purchased an unlimited number of times
- * @param excludeLimited Pass True to exclude gifts that can be purchased a limited number of times
+ * @param excludeLimitedUpgradable Pass True to exclude gifts that can be purchased a limited number of times and can be upgraded to unique
+ * @param excludeLimitedNonUpgradable Pass True to exclude gifts that can be purchased a limited number of times and can't be upgraded to unique
  * @param excludeUnique Pass True to exclude unique gifts
+ * @param excludeFromBlockchain Pass True to exclude gifts that were assigned from the TON blockchain and can't be resold or transferred in Telegram
  * @param sortByPrice Pass True to sort results by gift price instead of send date. Sorting is applied before pagination.
  * @param offset Offset of the first entry to return as received from the previous request; use empty string to get the first chunk of results
  * @param limit The maximum number of gifts to be returned; 1-100. Defaults to 100