+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/options/GetChatGiftsOptions.kt
@@ -3,15 +3,16 @@ package eu.vendeli.tgbot.types.options
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetBusinessAccountGiftsOptions(
+data class GetChatGiftsOptions(
     var excludeUnsaved: Boolean? = null,
     var excludeSaved: Boolean? = null,
     var excludeUnlimited: Boolean? = null,
     var excludeLimitedUpgradable: Boolean? = null,
     var excludeLimitedNonUpgradable: Boolean? = null,
-    var excludeUnique: Boolean? = null,
     var excludeFromBlockchain: Boolean? = null,
+    var excludeUnique: Boolean? = null,
     var sortByPrice: Boolean? = null,
     var offset: String? = null,
     var limit: Int? = null,
 ) : Options
+