+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/options/GetUserGiftsOptions.kt
@@ -3,15 +3,14 @@ package eu.vendeli.tgbot.types.options
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetBusinessAccountGiftsOptions(
-    var excludeUnsaved: Boolean? = null,
-    var excludeSaved: Boolean? = null,
+data class GetUserGiftsOptions(
     var excludeUnlimited: Boolean? = null,
     var excludeLimitedUpgradable: Boolean? = null,
     var excludeLimitedNonUpgradable: Boolean? = null,
-    var excludeUnique: Boolean? = null,
     var excludeFromBlockchain: Boolean? = null,
+    var excludeUnique: Boolean? = null,
     var sortByPrice: Boolean? = null,
     var offset: String? = null,
     var limit: Int? = null,
 ) : Options
+