+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/options/RepostStoryOptions.kt
@@ -1,0 +1,10 @@
+package eu.vendeli.tgbot.types.options
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RepostStoryOptions(
+    var postToChatPage: Boolean? = null,
+    var protectContent: Boolean? = null,
+) : Options
+