+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/options/SendMessageDraftOptions.kt
@@ -1,0 +1,8 @@
+package eu.vendeli.tgbot.types.options
+
+import eu.vendeli.tgbot.types.component.ParseMode
+
+data class SendMessageDraftOptions(
+    override var parseMode: ParseMode? = null,
+    override var messageThreadId: Int? = null,
+) : ForumProps, OptionsParseMode