+++ telegram-bot/src/jvmTest/kotlin/eu/vendeli/api/story/StoryApiTest.kt
@@ -4,6 +4,7 @@ import BotTestContext
 import eu.vendeli.tgbot.api.story.deleteStory
 import eu.vendeli.tgbot.api.story.editStory
 import eu.vendeli.tgbot.api.story.postStory
+import eu.vendeli.tgbot.api.story.repostStory
 import eu.vendeli.tgbot.types.component.ImplicitFile
 import eu.vendeli.tgbot.types.story.InputStoryContent
 import eu.vendeli.tgbot.utils.common.toImplicitFile
@@ -48,6 +49,15 @@ class StoryApiTest : BotTestContext() {
     @Test
     fun `deleteStory test`() = runTest {
         deleteStory("test", "story_id")
+            .sendReq()
+            .shouldFailure()
+            .description
+            .shouldBe("Bad Request: business connection not found")
+    }
+
+    @Test
+    fun `repostStory test`() = runTest {
+        repostStory("test", 123456L, 1, 86400.seconds)
             .sendReq()
             .shouldFailure()
             .description